daemon: move to account provider, use Iterator get_assets_for
