+++ xelis_daemon/src/core/storage/providers/nonce.rs
@@ -1,5 +1,5 @@
 use async_trait::async_trait;
-use log::{trace, error};
+use log::trace;
 use xelis_common::{
     account::VersionedNonce,
     block::TopoHeight,
@@ -11,10 +11,10 @@ use crate::core::{
     storage::{sled::ACCOUNTS_COUNT, SledStorage},
 };
 
-use super::{AssetProvider, BalanceProvider, NetworkProvider};
+use super::NetworkProvider;
 
 #[async_trait]
-pub trait NonceProvider: BalanceProvider {
+pub trait NonceProvider {
     // Check if the account has a nonce
     async fn has_nonce(&self, key: &PublicKey) -> Result<bool, BlockchainError>;
 
@@ -35,10 +35,6 @@ pub trait NonceProvider: BalanceProvider {
 
     // Get the nonce under or equal topoheight requested for an account
     async fn get_nonce_at_maximum_topoheight(&self, key: &PublicKey, topoheight: TopoHeight) -> Result<Option<(TopoHeight, VersionedNonce)>, BlockchainError>;
-
-    // Check if the account has a nonce updated in the range given
-    // It will also check balances if no nonce found
-    async fn has_key_updated_in_range(&self, key: &PublicKey, minimum_topoheight: TopoHeight, maximum_topoheight: TopoHeight) -> Result<bool, BlockchainError>;
 
     // Set the last topoheight that the account has a nonce changed
     async fn set_last_topoheight_for_nonce(&mut self, key: &PublicKey, topoheight: TopoHeight) -> Result<(), BlockchainError>;
@@ -169,83 +165,6 @@ impl NonceProvider for SledStorage {
         }
 
         Ok(None)
-    }
-
-    async fn has_key_updated_in_range(&self, key: &PublicKey, minimum_topoheight: TopoHeight, maximum_topoheight: TopoHeight) -> Result<bool, BlockchainError> {
-        trace!("has key {} updated in range min topoheight {} and max topoheight {}", key.as_address(self.is_mainnet()), minimum_topoheight, maximum_topoheight);
-        // check first that this address has nonce, if no returns None
-        if !self.has_nonce(key).await? {
-            return Ok(false)
-        }
-
-        // fast path check the latest nonce
-        let (topo, mut version) = self.get_last_nonce(key).await?;
-        trace!("Last version of nonce for {} is at topoheight {}", key.as_address(self.is_mainnet()), topo);
-
-        // if it's the latest and its under the maximum topoheight and above minimum topoheight
-        if topo >= minimum_topoheight && topo <= maximum_topoheight {
-            trace!("Last version nonce (valid) found at {} (maximum topoheight = {})", topo, maximum_topoheight);
-            return Ok(true)
-        }
-
-        // otherwise, we have to go through the whole chain
-        while let Some(previous) = version.get_previous_topoheight() {
-            // we are under the minimum topoheight, we can stop
-            if previous < minimum_topoheight {
-                break;
-            }
-
-            let previous_version = self.get_nonce_at_exact_topoheight(key, previous).await?;
-            trace!("previous nonce version is at {}", previous);
-            if previous <= maximum_topoheight {
-                trace!("Highest version nonce found at {} (maximum topoheight = {})", previous, maximum_topoheight);
-                return Ok(true)
-            }
-
-            // security in case of DB corruption
-            if let Some(value) = previous_version.get_previous_topoheight() {
-                if value > previous {
-                    error!("FATAL ERROR: Previous topoheight ({}) should not be higher than current version ({})!", value, previous);
-                    return Err(BlockchainError::Unknown)
-                }
-            }
-            version = previous_version;
-        }
-
-        // if we are here, we didn't find any nonce in the range
-        // it start to be more and more heavy...
-        // lets check on balances now...
-
-        // check that we have a VersionedBalance between range given
-        for asset in self.get_assets_for(key).await? {
-            let (topo, mut version) = self.get_last_balance(key, &asset).await?;
-            if topo >= minimum_topoheight && topo <= maximum_topoheight {
-                return Ok(true)
-            }
-
-            while let Some(previous) = version.get_previous_topoheight() {
-                // we are under the minimum topoheight, we can stop
-                if previous < minimum_topoheight {
-                    break;
-                }
-
-                let previous_version = self.get_balance_at_exact_topoheight(key, &asset, previous).await?;
-                if previous <= maximum_topoheight {
-                    return Ok(true)
-                }
-
-                // security in case of DB corruption
-                if let Some(value) = previous_version.get_previous_topoheight() {
-                    if value > previous {
-                        error!("FATAL ERROR: Previous topoheight for balance ({}) should not be higher than current version of balance ({})!", value, previous);
-                        return Err(BlockchainError::Unknown)
-                    }
-                }
-                version = previous_version;
-            }
-        }
-
-        Ok(false)
     }
 
     async fn set_nonce_at_topoheight(&mut self, key: &PublicKey, topoheight: TopoHeight, version: &VersionedNonce) -> Result<(), BlockchainError> {
